requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
bcrypt==4.2.1
python-multipart==0.0.20
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
